refactor: 拆分灵动岛鼠标状态机和颜色时间hooks
